requirements.txt
discord.py>=2.3.2
aiohttp>=3.9.1
pyyaml>=6.0.1
python-dateutil>=2.8.2
humanize>=4.9.0